chore: absorve code stream to separated file
